merge

+++ js/game.js
@@ -19,12 +19,13 @@
 		{
 			antialias: true,
 			stencil: false,
-			precision: "lowp",
+			precision: "highp",
 			preserveDrawingBuffer: false,
 			depth: true
 		});
 
 		this.camera = new Camera(45,  window.innerWidth / window.innerHeight, 0.1, 10000);
+
 		this.stationFloor = new StationFloor()
 		this.sun = new Sun();
 		this.hemisphere = new Hemisphere();
